refactor: header hook 분리 #203
